seperated registry to a new file
